api/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.2
pillow==11.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
